Major blueprint migration: Move project and task routes
Projects Blueprint:
- Move view_project and edit_project routes from app.py
- Add proper error handling and access control
- Update URL redirects to use blueprint namespaces

Tasks Blueprint:
- Move main tasks list route with complex filtering logic
- Move task deletion route with proper access control
- Implement comprehensive task filtering (status, priority, assignee, project, date)
- Support for both project tasks and independent tasks
- Handle interdependent project task sequencing

Progress: Now moved 11+ routes from app.py to appropriate blueprints.
Systematic blueprint refactoring continues with improved code organization.
