+++ blueprints/projects.py
@@ -5,7 +5,7 @@ Project management blueprint
 from flask import Blueprint, render_template, request, redirect, url_for, session, flash, jsonify
 from datetime import datetime
 from core import db
-from models import Project, Template, Task, Client, TaskStatus, TemplateTask, WorkType
+from models import Project, Template, Task, Client, TaskStatus, TemplateTask, WorkType, User, ActivityLog
 from utils import calculate_task_due_date, find_or_create_client, create_activity_log
 
 projects_bp = Blueprint('projects', __name__, url_prefix='/projects')
@@ -118,3 +118,55 @@ def create_project():
     templates = Template.query.filter_by(firm_id=firm_id).all()
     clients = Client.query.filter_by(firm_id=firm_id, is_active=True).all()
     return render_template('projects/create_project.html', templates=templates, clients=clients)
+
+
+@projects_bp.route('/<int:id>')
+def view_project(id):
+    project = Project.query.get_or_404(id)
+    if project.firm_id != session['firm_id']:
+        flash('Access denied', 'error')
+        return redirect(url_for('projects.list_projects'))
+    
+    tasks = Task.query.filter_by(project_id=id).order_by(Task.due_date.asc()).all()
+    activity_logs = ActivityLog.query.filter_by(project_id=id).order_by(ActivityLog.timestamp.desc()).limit(10).all()
+    
+    return render_template('projects/view_project.html', project=project, tasks=tasks, activity_logs=activity_logs)
+
+
+@projects_bp.route('/<int:id>/edit', methods=['GET', 'POST'])
+def edit_project(id):
+    project = Project.query.get_or_404(id)
+    if project.firm_id != session['firm_id']:
+        flash('Access denied', 'error')
+        return redirect(url_for('projects.list_projects'))
+    
+    if request.method == 'POST':
+        # Update project details
+        project.name = request.form.get('name')
+        project.priority = request.form.get('priority', 'Medium')
+        project.status = request.form.get('status', 'Active')
+        
+        # Handle due date
+        due_date = request.form.get('due_date')
+        if due_date:
+            project.due_date = datetime.strptime(due_date, '%Y-%m-%d').date()
+        else:
+            project.due_date = None
+        
+        # Handle start date
+        start_date = request.form.get('start_date')
+        if start_date:
+            project.start_date = datetime.strptime(start_date, '%Y-%m-%d').date()
+        
+        db.session.commit()
+        
+        # Activity log
+        create_activity_log(f'Project "{project.name}" updated', session['user_id'], project.id)
+        
+        flash('Project updated successfully!', 'success')
+        return redirect(url_for('projects.view_project', id=project.id))
+    
+    # GET request - show form
+    firm_id = session['firm_id']
+    users = User.query.filter_by(firm_id=firm_id).all()
+    return render_template('projects/edit_project.html', project=project, users=users)
